Splitt FrontendApi og MaskinApi
